+++ src/lib.rs
@@ -1,8 +1,4 @@
-use std::{collections::HashMap, sync::Arc};
-
 use raxb::quick_xml::events::Event;
-
-pub use xoev_xwasser_codelists::{CodeList, XWasserCodeListValue};
 
 #[cfg(feature = "builder")]
 pub mod builder;
@@ -10,8 +6,13 @@ pub mod model;
 #[cfg(not(feature = "wasm"))]
 #[cfg(feature = "schema")]
 pub mod schemas;
+mod validate;
 #[cfg(feature = "wasm")]
 pub mod wasm;
+
+pub use xoev_xwasser_codelists::XWasserCodeListValue;
+
+pub use validate::{XWasserValidate, XWasserValidateError, XWasserValidateMarker};
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum Version {
@@ -74,65 +75,4 @@ pub fn detect_version(xml: &str) -> Version {
         }
     }
     Version::Unknown
-}
-
-pub trait XWasserValidate {
-    fn xwasser_validate(
-        &self,
-        codelists: &HashMap<Arc<str>, CodeList>,
-    ) -> Result<(), XWasserValidateError>;
-}
-
-pub enum XWasserValidateError {
-    CodeListValueNotFound { codelist: String, value: String },
-}
-
-pub trait XWasserValidateMarker {}
-
-impl<T> XWasserValidate for T
-where
-    T: XWasserCodeListValue + XWasserValidateMarker,
-{
-    fn xwasser_validate(
-        &self,
-        codelists: &HashMap<Arc<str>, CodeList>,
-    ) -> Result<(), XWasserValidateError> {
-        if self.validate(codelists) {
-            Ok(())
-        } else {
-            Err(XWasserValidateError::CodeListValueNotFound {
-                codelist: Self::CODELIST.to_string(),
-                value: self.as_value().to_string(),
-            })
-        }
-    }
-}
-
-impl<T> XWasserValidate for Option<T>
-where
-    T: XWasserValidate,
-{
-    fn xwasser_validate(
-        &self,
-        codelists: &HashMap<Arc<str>, CodeList>,
-    ) -> Result<(), XWasserValidateError> {
-        self.as_ref()
-            .map(|t| t.xwasser_validate(codelists))
-            .unwrap_or(Ok(()))
-    }
-}
-
-impl<T> XWasserValidate for Vec<T>
-where
-    T: XWasserValidate,
-{
-    fn xwasser_validate(
-        &self,
-        codelists: &HashMap<Arc<str>, CodeList>,
-    ) -> Result<(), XWasserValidateError> {
-        self.iter()
-            .find_map(|t| t.xwasser_validate(codelists).err())
-            .map(Err)
-            .unwrap_or(Ok(()))
-    }
 }
